refactor: create NodeManager and remove Node.from_token
