+++ tests/unit_tests/test_node.py
@@ -1,165 +1,22 @@
 import pytest
 
-from sqlfmt.exception import SqlfmtBracketError
 from sqlfmt.mode import Mode
-from sqlfmt.node import Node
+from sqlfmt.node_manager import NodeManager
 from sqlfmt.token import Token, TokenType
 from tests.util import read_test_data
 
 
-def test_calculate_depth() -> None:
-    select_t = Token(
-        type=TokenType.UNTERM_KEYWORD,
-        prefix="",
-        token="select",
-        spos=0,
-        epos=6,
-    )
-    select_n = Node.from_token(token=select_t, previous_node=None)
-    assert (select_n.depth, select_n.open_brackets) == ((0, 0), [])
-
-    open_paren_t = Token(
-        type=TokenType.BRACKET_OPEN,
-        prefix="    ",
-        token="(",
-        spos=8,
-        epos=9,
-    )
-    open_paren_n = Node.from_token(token=open_paren_t, previous_node=select_n)
-    assert (open_paren_n.depth, open_paren_n.open_brackets) == ((1, 0), [select_n])
-
-    one_t = Token(
-        type=TokenType.NUMBER,
-        prefix=" ",
-        token="1",
-        spos=10,
-        epos=11,
-    )
-    one_n = Node.from_token(token=one_t, previous_node=open_paren_n)
-    assert (one_n.depth, one_n.open_brackets) == ((2, 0), [select_n, open_paren_n])
-
-    close_paren_t = Token(
-        type=TokenType.BRACKET_CLOSE,
-        prefix="",
-        token=")",
-        spos=11,
-        epos=12,
-    )
-    close_paren_n = Node.from_token(token=close_paren_t, previous_node=one_n)
-    assert (close_paren_n.depth, close_paren_n.open_brackets) == ((1, 0), [select_n])
-
-
-def test_calculate_depth_exception() -> None:
-    close_paren = Token(
-        type=TokenType.BRACKET_CLOSE,
-        prefix="",
-        token=")",
-        spos=0,
-        epos=1,
-    )
-    with pytest.raises(SqlfmtBracketError):
-        _ = Node.from_token(close_paren, previous_node=None)
-
-
-def test_jinja_depth(default_mode: Mode) -> None:
-    source_string, _ = read_test_data("unit_tests/test_node/test_jinja_depth.sql")
-    q = default_mode.dialect.initialize_analyzer(
-        line_length=default_mode.line_length
-    ).parse_query(source_string=source_string)
-    expected = [
-        (0, 0),  # {{ config(materialized="table") }}
-        (0, 0),  # \n
-        (0, 0),  # \n
-        (0, 0),  # {%- set n = 5 -%}
-        (0, 0),  # \n
-        (0, 0),  # with
-        (1, 0),  # \n
-        (1, 0),  # {% for i in range(n) %}
-        (1, 1),  # \n
-        (1, 1),  # dont_do_this_
-        (1, 1),  # {{ i }}
-        (1, 1),  # as
-        (1, 1),  # (
-        (2, 1),  # \n
-        (2, 1),  # {% if foo %}
-        (2, 2),  # \n
-        (2, 2),  # select
-        (3, 2),  # \n
-        (2, 1),  # {% elif bar %}
-        (2, 2),  # \n
-        (2, 2),  # select distinct
-        (3, 2),  # \n
-        (2, 1),  # {% elif baz %}
-        (2, 2),  # \n
-        (2, 2),  # select top 25
-        (3, 2),  # \n
-        (2, 1),  # {% else %}
-        (2, 2),  # \n
-        (2, 2),  # select
-        (3, 2),  # \n
-        (3, 1),  # {% endif %}
-        (3, 1),  # \n
-        (3, 1),  # my_col
-        (3, 1),  # \n
-        (2, 1),  # from
-        (3, 1),  # \n
-        (3, 1),  # {% if i == qux %}
-        (3, 2),  # \n
-        (3, 2),  # zip
-        (3, 2),  # \n
-        (3, 1),  # {% else %}
-        (3, 2),  # \n
-        (3, 2),  # zap
-        (3, 2),  # \n
-        (3, 1),  # {% endif %}
-        (3, 1),  # \n
-        (1, 1),  # )
-        (1, 1),  # {% if not loop.last %}
-        (1, 2),  # ,
-        (1, 1),  # {% endif%}
-        (1, 1),  # \n
-        (1, 0),  # {% endfor %}
-        (1, 0),  # \n
-        (1, 0),  # {% for i in range(n) %}
-        (1, 1),  # \n
-        (0, 1),  # select
-        (1, 1),  # \n
-        (1, 1),  # *
-        (1, 1),  # \n
-        (0, 1),  # from
-        (1, 1),  # \n
-        (1, 1),  # dont_do_this_
-        (1, 1),  # {{ i }}
-        (1, 1),  # \n
-        (1, 1),  # {% if not loop.last -%}
-        (1, 2),  # \n
-        (0, 2),  # union all
-        (0, 2),  # \n
-        (0, 1),  # {%- endif %}
-        (0, 1),  # \n
-        (0, 0),  # {% endfor %}
-        (0, 0),  # \n
-    ]
-    actual = [node.depth for node in q.nodes]
-    assert actual == expected
-
-
-def test_from_token_raises_bracket_error_on_jinja_block_end() -> None:
-    t = Token(
-        type=TokenType.JINJA_BLOCK_END,
-        prefix="",
-        token="{% endif %}",
-        spos=0,
-        epos=11,
-    )
-    with pytest.raises(SqlfmtBracketError):
-        _ = Node.from_token(t, previous_node=None)
+@pytest.fixture
+def node_manager(default_mode: Mode) -> NodeManager:
+    return NodeManager(default_mode.dialect.case_sensitive_names)
 
 
 @pytest.mark.parametrize(
     "token,result", [("between", True), ("BETWEEN", True), ("like", False)]
 )
-def test_is_the_between_operator(token: str, result: bool) -> None:
+def test_is_the_between_operator(
+    token: str, result: bool, node_manager: NodeManager
+) -> None:
     t = Token(
         type=TokenType.WORD_OPERATOR,
         prefix="",
@@ -167,14 +24,16 @@ def test_is_the_between_operator(token: str, result: bool) -> None:
         spos=0,
         epos=7,
     )
-    n = Node.from_token(t, previous_node=None)
+    n = node_manager.create_node(t, previous_node=None)
     assert n.is_the_between_operator is result
 
 
 @pytest.mark.parametrize(
     "token,result", [("[", True), ("(", False), ("]", False), ("{", False)]
 )
-def test_is_opening_square_bracket(token: str, result: bool) -> None:
+def test_is_opening_square_bracket(
+    token: str, result: bool, node_manager: NodeManager
+) -> None:
     t = Token(
         type=TokenType.BRACKET_OPEN,
         prefix="",
@@ -182,7 +41,7 @@ def test_is_opening_square_bracket(token: str, result: bool) -> None:
         spos=0,
         epos=1,
     )
-    n = Node.from_token(t, previous_node=None)
+    n = node_manager.create_node(t, previous_node=None)
     assert n.is_opening_square_bracket is result
 
 
@@ -205,36 +64,3 @@ def test_is_the_and_after_the_between_operator(default_mode: Mode) -> None:
     assert all([not n.is_the_and_after_the_between_operator for n in boolean_ands])
     assert all([n.is_the_and_after_the_between_operator for n in between_ands])
     assert all([not n.is_the_and_after_the_between_operator for n in other_nodes])
-
-
-def test_union_depth(default_mode: Mode) -> None:
-    source_string, _ = read_test_data("unit_tests/test_node/test_union_depth.sql")
-    q = default_mode.dialect.initialize_analyzer(
-        line_length=default_mode.line_length
-    ).parse_query(source_string=source_string)
-
-    expected = [
-        (0, 0),  # select,
-        (1, 0),  # 1,
-        (1, 0),  # \n,
-        (1, 0),  # \n,
-        (0, 0),  # union,
-        (0, 0),  # \n,
-        (0, 0),  # \n,
-        (0, 0),  # select,
-        (1, 0),  # 2,
-        (1, 0),  # \n,
-        (1, 0),  # \n,
-        (0, 0),  # union all,
-        (0, 0),  # \n,
-        (0, 0),  # \n,
-        (0, 0),  # (,
-        (1, 0),  # \n,
-        (1, 0),  # select,
-        (2, 0),  # 3,
-        (2, 0),  # \n,
-        (0, 0),  # ),
-        (0, 0),  # \n'
-    ]
-    actual_depth = [n.depth for n in q.nodes]
-    assert actual_depth == expected
